Trap Handling, Move List, Test Cleanup
Added code to process the moves.txt file and store the parsed Moves as a static object, and to fetch existing ones by name.

Updated testing to utilize the above.

Reordered some trait names in the moves list to be more consistent.

Added the option for negation in the WHEN clause for Effect parsing.

Added the TOTAL stat, a number that is added to the final damage total after all hits are calculated.

Updated the SUB operation so that it can go into negatives. This was to make sure the number of dice rolled and hit per die couldn't be negative, but was preventing the TOTAL stat from being reduced properly.