Game Flow, First Pass
Implemented processRound() method for the Duel class, which will
facilitate each individual turn once moves are chosen. Determining
player order, and processing a move by first applying all relevant
effects are the primary effects.

Changed Player class to contain the last move played rather than
referencing it as a static object in the Duel class. Round processing
acts by changing the "active" and "opposing" Player as it handles the
first and second move. Additionally modified the addHealth() method of
the Player class so that it properly bounds to 0 and the baseHealth.

Added a process() method to the Move class to handle the finer details
of dice rolling, damage dealing, and things of that nature. Also added a
getSpeed() method to convert whether or not a move is FAST/SLOW/neither
into an integer value for easy comparison.

Created a dedicated RNGesus class for all randomized actions, which for
now are just rolling dice and deciding who goes first in the event of a
tie.

Created a Tester class. Moved all tests out of the MoveParser main
method and into this.