The Great Big IRC Commit
Updated the Printable trait to be more like how I implemented it in
HistoryEngine.

Added a SET: clause to the EffectParser so that effects can update flags
for the effect. Needed by Ice Wall so that the healing effect isn't
marked as a TRAP, which was causing it to be removed before end of turn
previously.

Updated a lot of print statements to be more appropriate to game
narration.

Created the GameState enumeration to track the status of the current
Duel.

Filled out the game/round/move logic in the Duel class to facilitate
handling each round, one at a time, and parsing effects when
appropriate.

Filled out the DiceLegend class with heaps of code to handle individual
commands from users and the logic it uses. There is a hell of a lot
going on here and I need to document it soon.

Updated the Ice Wall move definition to use the new SET: clause, and to
correctly look for stacks of the BlessedFountain effect.