Elemental Dummy
Added the option to specify a particular Trait for the dummy to
exclusively use. I intend for people to choose one of the 4 elements
when using this, but the logic allows for any Trait if desired.

As part of this, moved the random-move-choice call to a distinct
function, as it is now more complex than a single assignment operation.