Case Sensitivity, Trait null-checking
Updated the static get() method of the Trait class, and the static
getMove() method of the MoveParser class to both now be case
insensitive.

Changed the startDuel() method of the Duel class, and the
startPractice() method of the IrcBot class to no longer default to null
for the element parameter. Instead, it uses the new NONE object in the
Trait class, which evaluates to false when checked for truthiness.