Slash Commands: Initial Go
Created a Command class to handle the 4 elements of a slash command: the
name that you actually type, the description provided to the user,
additional actions taken to set up the command, and the actions taken
when the command is executed.

This works as expected. It took a change in permissions on the Discord
side, however -- make sure that slash commands are enabled in the bot
permissions when you generate the URL that adds it to your server.