Discord: Second Pass, Functions
Created a Command class that contains the data for generating a slash command in Discord. It contains the name of the command, the description, additional steps to take in setup, and actions to take when the command is executed.

This is mostly functional -- the /move command still needs to be implemented, but once it is, this may be 'done' in the sense that it will mirror IRC functionality. I'd like to make move and practice-element choice either auto-fill or have buttons to choose, eventually.