Discord: Move Command
Set up the /move command, works as intended.

Modified the print() call for the DiscordBot class to no longer send
each message one at a time, but instead pile it into a single message
that is sent all at once when the moves are done being processed. We
were being rate-limited prior to this.